Add authentification to routes when possible

+++ controllers/user.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"projet_wiki/database"
 	"projet_wiki/models"
-
+	"projet_wiki/controllers/auth"
 	"github.com/gorilla/mux"
 	"golang.org/x/crypto/bcrypt"
 )
@@ -65,6 +65,15 @@ func DeleteUser(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		fmt.Println(err)
 	}
+	UserAuth, errAuth := auth.GetRequestUser(r)
+	if errAuth != nil {
+		fmt.Println("user not connected")
+		return
+	}
+	if UserAuth.Username != user.Username {
+		fmt.Println("Incorrect account rights ")
+	    return
+	}
 	db.Delete(&user)
 }
 
@@ -83,6 +92,16 @@ func UpdateUser(w http.ResponseWriter, r *http.Request) {
 		fmt.Println(err)
 	}
 
+	UserAuth, errAuth := auth.GetRequestUser(r)
+	if errAuth != nil {
+		fmt.Println("user not connected")
+		return
+	}
+	if UserAuth.Username != user.Username {
+		fmt.Println("Incorrect account rights ")
+	    return
+	}
+	
 	user.Username = username
 
 	err2 := db.Save(&user)
